web_endpoint/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.txt pyproject.toml
#
altair==5.0.1
    # via streamlit
attrs==23.1.0
    # via jsonschema
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
blinker==1.6.2
    # via streamlit
boto3==1.26.146
    # via yt-trending-mlops (pyproject.toml)
botocore==1.29.146
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.1
    # via streamlit
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via streamlit
decorator==5.1.1
    # via validators
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==6.6.0
    # via streamlit
importlib-resources==5.12.0
    # via jsonschema
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via altair
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.24.3
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==23.1
    # via streamlit
pandas==2.0.2
    # via
    #   altair
    #   streamlit
    #   yt-trending-mlops (pyproject.toml)
pillow==9.5.0
    # via streamlit
pkgutil-resolve-name==1.3.10
    # via jsonschema
protobuf==4.23.2
    # via streamlit
pyarrow==12.0.0
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
    #   streamlit
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.31.0
    # via
    #   streamlit
    #   yt-trending-mlops (pyproject.toml)
rich==13.4.1
    # via streamlit
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
streamlit==1.23.1
    # via yt-trending-mlops (pyproject.toml)
tenacity==8.2.2
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via streamlit
typing-extensions==4.6.3
    # via
    #   altair
    #   rich
    #   streamlit
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
urllib3==1.26.16
    # via
    #   botocore
    #   requests
validators==0.20.0
    # via streamlit
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources
